requirements/base.txt
Fraction~=2.2.0
Pillow~=10.1.0  # https://github.com/python-pillow/Pillow
argon2-cffi~=23.1.0  # https://github.com/hynek/argon2_cffi
coveralls~=3.3.1
django-allauth~=0.58.0  # https://github.com/pennersr/django-allauth
django-anymail[mailgun]~=10.2  # https://github.com/anymail/django-anymail
django-compressor~=4.4  # https://github.com/django-compressor/django-compressor
django-cors-headers~=4.3.0 # https://github.com/adamchainz/django-cors-headers
django-crispy-forms~=2.1 # https://github.com/django-crispy-forms/django-crispy-forms
django-environ~=0.11.2  # https://github.com/joke2k/django-environ
django-filter>=2.2,<3.0 # this constraint is required for wagtail
django-mjml[requests]~=1.1
django-taggit~=4.0.0
django-modelcluster~=6.0
django-model-utils~=4.3.0  # https://github.com/jazzband/django-model-utils
django-oauth-toolkit~=2.2.0
django-phonenumber-field~=7.0.2
django-storages[google]~=1.13.2  # https://github.com/jschneier/django-storages
djangorestframework-datatables~=0.7.0
djangorestframework~=3.14.0  # https://github.com/encode/django-rest-framework
django-rest-framework-braces~=0.3.4
drf-generators~=0.5.0
django~=4.2  # pyup: < 3.2  # https://www.djangoproject.com/
google-api-core~=2.11.0
google-api-python-client~=2.86.0
google-auth~=2.17.3
google-cloud-secret-manager~=2.16.1
google-cloud-storage~=2.8.0
gcloud~=0.18.3
protobuf~=4.22.3
openpyxl~=3.1.2
phonenumbers~=8.13.10
python-slugify~=8.0.1  # https://github.com/un33k/python-slugify
pytz~=2023.3  # https://github.com/stub42/pytz
rcssmin~=1.1.1  # https://github.com/ndparker/rcssmin
requests~=2.28.2
uvicorn[standard]~=0.21.1  # https://github.com/encode/uvicorn
whitenoise~=6.6.0  # https://github.com/evansd/whitenoise
wagtail~=5.1
Wand~=0.6.11
opencv-python==4.7.0.72
rustface~=0.1.0
wagtailvideos~=4.2.0
wagtailmedia~=0.14.0
django-nested-admin==4.0.2
